requirements.txt
python-dotenv
heyoo
fastapi
uvicorn
tinydb
google-generativeai
gspread
gspread_dataframe
oauth2client